+++ system/Cache/Handlers/MemcachedHandler.php
@@ -115,8 +115,6 @@ class MemcachedHandler extends BaseHandler
             } else {
                 throw new CriticalError('Cache: Not support Memcache(d) extension.');
             }
-        } catch (CriticalError $e) {
-            throw $e;
         } catch (Exception $e) {
             throw new CriticalError('Cache: Memcache(d) connection refused (' . $e->getMessage() . ').');
         }