+++ tests/system/Database/Builder/InsertTest.php
@@ -64,7 +64,7 @@ final class InsertTest extends CIUnitTestCase
         $builder = $this->db->table('jobs');
 
         $this->expectException(DatabaseException::class);
-        $this->expectExceptionMessage('You must use the "set" method to update an entry.');
+        $this->expectExceptionMessage('You must use the "set" method to insert an entry.');
 
         $builder->testMode()->insert(null, true);
     }