+++ user_guide_src/source/outgoing/view_decorators/001.php
@@ -2,7 +2,7 @@
 
 namespace App\Views\Decorators;
 
-use CodeIgniter\Views\ViewDecoratorInterface;
+use CodeIgniter\View\ViewDecoratorInterface;
 
 class MyDecorator implements ViewDecoratorInterface
 {